+++ server/src/main/java/me/deit/server/security/AuthController.java
@@ -1,10 +1,7 @@
-package me.deit.server.controllers;
+package me.deit.server.security;
 
-import me.deit.server.repositories.UserRepository;
 import me.deit.server.security.jwt.JwtUtils;
-import me.deit.server.services.JwtResponse;
-import me.deit.server.services.LoginRequest;
-import me.deit.server.services.UserDetailsImpl;
+import me.deit.server.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -15,8 +12,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
