+++ server/src/main/java/me/deit/server/security/UserDetailsImpl.java
@@ -1,60 +1,54 @@
-package me.deit.server.entities;
+package me.deit.server.security;
 
-import javax.persistence.*;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import me.deit.server.user.User;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 
-@Entity
-@Table(name="user", schema = "public",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = "email")
-        })
-public class User {
+import java.util.Collection;
 
-    @Id
-    @Column (name = "id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//new class implementing UserDetails and ready for new business logic
+
+public class UserDetailsImpl implements UserDetails {
+
     private Long id;
-
-    @Column (name = "email")
-    @Email
     private String email;
 
-    @Column (name= "password")
+    @JsonIgnore
     private String password;
 
-    @Column (name = "first_name")
     private String first_name;
-
-    @Column (name = "last_name")
     private String last_name;
-
-    @Column (name = "phone_number")
     private String phone_number;
-
-    @Column (name = "gender")
     private String gender;
-
-    @Column (name = "preference")
     private String preference;
-
-    @Column (name = "description")
     private String description;
 
-    public User() {
-    }
-
-    public User(Long id, @Email String email, String password, @NotBlank String first_name, @NotBlank String last_name,
-                @NotBlank String phone_number, @NotBlank String gender, @NotBlank String preference, String description) {
+    public UserDetailsImpl(Long id, String email, String password, String first_name, String last_name, String phone_number, String gender,
+                           String preference, String description) {
         this.id = id;
         this.email = email;
         this.password = password;
         this.first_name = first_name;
         this.last_name = last_name;
-        this.phone_number = phone_number;
         this.gender = gender;
+        this.phone_number = phone_number;
         this.preference = preference;
         this.description = description;
+    }
+
+    public static UserDetailsImpl build(User user) {
+        return new UserDetailsImpl(
+                user.getId(),
+                user.getEmail(),
+                user.getPassword(),
+                user.getFirst_name(),
+                user.getLast_name(),
+                user.getPhone_number(),
+                user.getGender(),
+                user.getPreference(),
+                user.getDescription()
+        );
     }
 
     public Long getId() {
@@ -71,10 +65,6 @@ public class User {
 
     public void setEmail(String email) {
         this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
     }
 
     public void setPassword(String password) {
@@ -127,5 +117,41 @@ public class User {
 
     public void setDescription(String description) {
         this.description = description;
+    }
+
+    //roles?
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
+
+    @Override
+    public String getPassword() {
+        return password;
+    }
+
+    @Override
+    public String getUsername() {
+        return email;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
     }
 }
